+++ webApp/src/features/Modal/GlobalModals.tsx
@@ -18,6 +18,7 @@ import { DailyQuestionModal } from '@/features/DailyQuestion/DailyQuestionModal'
 import { GrammarImprovementModal } from '@/features/Dashboard/Grammar/GrammarImprovementModal';
 import { ProgressStatModal } from '@/features/ProgressStat/ProgressStatModal';
 import { EssayModal } from '../Essay/EssayModal';
+import { NewsModal } from '@/features/News/NewsModal';
 
 export const GlobalModals: React.FC = () => {
   const game = useGame();
@@ -71,6 +72,7 @@ export const GlobalModals: React.FC = () => {
       <TeacherVoiceModal />
 
       <GrammarImprovementModal />
+      <NewsModal />
     </>
   );
 };